requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
coinapi.rest.v1==1.3
colorama==0.4.4
discord.py==1.7.3
docutils==0.16
idna==2.10
importlib-metadata==3.7.2
keyring==23.0.0
multidict==6.0.2
numpy==1.22.2
packaging==20.9
pandas==1.4.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
soupsieve==2.3.1
TableauScraper==0.1.29
tqdm==4.59.0
twine==3.3.0
typing_extensions==4.0.1
urllib3==1.26.3
webencodings==0.5.1
wincertstore==0.2
yarl==1.7.2
zipp==3.4.1
